Backend/requirements.txt
flask==2.3.2
openai==0.27.8
python-dotenv==1.0.0
flask-cors==3.0.10
flask
twilio
python-dotenv
flask-cors
flask-socketio==5.5.1
eventlet==0.40.3
